requirements.txt
asgiref==3.8.1
autopep8==2.3.2
Django==5.1.5
django-filter==24.3
django-silk==5.3.2
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
Faker==33.3.1
gprof2dot==2024.6.6
isort==6.0.0
pillow==11.1.0
psycopg2-binary==2.9.10
pycodestyle==2.12.1
PyJWT==2.10.1
python-dateutil==2.9.0.post0
six==1.17.0
sqlparse==0.5.3
tomli==2.2.1
typing_extensions==4.12.2
